backend/requirements.txt
Flask==2.0.1
pandas==1.3.3
numpy==1.21.2
openpyxl==3.0.9
fpdf2==2.5.5
plotly==5.10.0